+++ src/libc/dos/compat/d_read.c
@@ -17,6 +17,7 @@
 #include <go32.h>
 #include <dpmi.h>
 #include <errno.h>
+#include <stdint.h>
 #include <dos.h>
 
 unsigned int _dos_read(int handle, void *buffer, unsigned int count, unsigned int *result)
@@ -54,7 +55,7 @@ unsigned int _dos_read(int handle, void *buffer, unsigned int count, unsigned in
       return r.x.ax;
     }
     if ( r.x.ax )
-      movedata(dos_selector, 0, _my_ds(), (unsigned int)p_buffer, r.x.ax);
+      movedata(dos_selector, 0, _my_ds(), (uintptr_t)p_buffer, r.x.ax);
     count    -= read_size;
     p_buffer += r.x.ax;
     *result  += r.x.ax;
